feat: implement -t, -r options, refine quick_sort algorithm

+++ libft/ft_printf/ft_printf.h
@@ -23,7 +23,7 @@
 typedef struct s_printf
 {
 	char			flag;
-	int				width;
+	size_t			width;
 	size_t			prec;
 	int				prec_flag;
 	char			type;
